requirements.txt
# ===== 核心框架 =====
Django>=4.2,<5.0
Pillow>=10.0.0
requests>=2.31.0

# ===== AI / LLM =====
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0

# ===== 计算机视觉 =====
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0

# ===== GroundingDINO 依赖 =====
transformers>=4.36.0,<5.0.0
timm>=0.9.0
addict>=2.4.0
yapf>=0.40.0
supervision>=0.22.0
pycocotools>=2.0.0

# ===== 时序预测 =====
prophet>=1.1.0
pandas>=2.0.0

# ===== 工具库 =====
scipy>=1.10.0
huggingface_hub>=0.20.0